requirements.txt
Django==4.2
psycopg2-binary==2.9.7
python-dotenv==1.0.1
djangorestframework==3.14.0
pytest==8.0.1
pytest-django==4.8.0
geopy==2.4.1
redis==5.0.2
celery==5.3.6